pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aor-runtime"
version = "0.1.0"
description = "Local-first agent orchestration runtime built on LangGraph."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116,<1.0",
  "langchain-openai>=1.0.0,<2.0.0",
  "langgraph>=1.1.6,<2.0.0",
  "pydantic>=2.11,<3.0",
  "pyyaml>=6.0.2,<7.0",
  "rich>=14.0,<15.0",
  "sqlalchemy>=2.0.40,<3.0",
  "typer>=0.16,<1.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "ipython>=9.0,<10.0",
]

[project.scripts]
aor = "aor_runtime.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
